+++ src/roswire/ros1/node.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-__all__ = ('Node',)
+__all__ = ('ROS1Node',)
 
 from typing import Optional
 import xmlrpc.client
@@ -8,21 +8,18 @@ import dockerblade
 import psutil
 
 from ..exceptions import ROSWireException
+from ..interface import Node
 
 
-class Node:
+class ROS1Node(Node):
     """Provides access to a ROS node.
 
     Attributes
     ----------
-    api: xmlrpc.client.ServerProxy
-        An XML-RPC API client for this node.
     name: str
         The fully qualified name of this node.
-    url: str
-        URL used to access this node from the host network.
-    pid: int
-        The container PID of the main process for this node.
+    _url: str
+        The URL used to access this node from the host network.
     pid_host: int
         The host PID of the main process for this node.
     """
@@ -32,24 +29,22 @@ class Node:
                  shell: dockerblade.Shell
                  ) -> None:
         self.__name = name
-        self.__url = url_host_network
+        self._url = url_host_network
         self.__shell = shell
         self.__pid_host: Optional[int] = None
-
-    @property
-    def api(self) -> xmlrpc.client.ServerProxy:
-        return xmlrpc.client.ServerProxy(self.url)
 
     @property
     def name(self) -> str:
         return self.__name
 
     @property
-    def url(self) -> str:
-        return self.__url
+    def _api(self) -> xmlrpc.client.ServerProxy:
+        """An XML-RPC API client for interacting with this node."""
+        return xmlrpc.client.ServerProxy(self._url)
 
     @property
-    def pid(self) -> int:
+    def _pid(self) -> int:
+        """The container PID of the main process for this node."""
         code, status, pid = self.api.getPid('/.roswire')  # type: ignore
         if code != 1:
             m = f"failed to obtain PID [{self.name}]: {status} (code: {code})"
@@ -59,9 +54,10 @@ class Node:
         return pid
 
     @property
-    def pid_host(self) -> int:
+    def _pid_host(self) -> int:
+        """The host PID of the main process for this node."""
         if self.__pid_host is None:
-            self.__pid_host = self.__shell._local_to_host_pid(self.pid)
+            self.__pid_host = self.__shell._local_to_host_pid(self._pid)
             assert self.__pid_host is not None
         return self.__pid_host
 
@@ -69,7 +65,7 @@ class Node:
         """Determines whether this node is alive."""
         # TODO check start time to ensure this is the same process!
         try:
-            return psutil.pid_exists(self.pid_host)
+            return psutil.pid_exists(self._pid_host)
         except ROSWireException:
             return False